Use Vuex store in tests
