+++ kolibri/plugins/management/assets/src/management.js
@@ -4,7 +4,7 @@ import logging from 'loglevel';
 import KolibriModule from 'kolibri_module';
 import Vue from 'vue';
 import Main from './Main.vue';
-import store from './vuex/store.js';
+import { store } from './vuex/store.js';
 
 const vm = new Vue({
   el: 'body',
